move printing candidate to core
+ expose `printCandidate` and `printVariant` on the DesignSystem
