+++ packages/@tailwindcss-upgrade/src/codemods/template/candidates.test.ts
@@ -1,7 +1,7 @@
 import { __unstable__loadDesignSystem } from '@tailwindcss/node'
 import { describe, expect, test } from 'vitest'
 import { spliceChangesIntoString } from '../../utils/splice-changes-into-string'
-import { extractRawCandidates, printCandidate } from './candidates'
+import { extractRawCandidates } from './candidates'
 
 let html = String.raw
 
@@ -190,7 +190,7 @@ describe('printCandidate()', () => {
 
     // Sometimes we will have a functional and a static candidate for the same
     // raw input string (e.g. `-inset-full`). Dedupe in this case.
-    let cleaned = new Set([...candidates].map((c) => printCandidate(designSystem, c)))
+    let cleaned = new Set([...candidates].map((c) => designSystem.printCandidate(c)))
 
     expect([...cleaned]).toEqual([result])
   })
